# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API is used to interact with Carbonaut resources",
        "title": "Carbonaut API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/v1/data/import/csv": {
            "post": {
                "description": "Import csv raw bytes of provider data to carbonaut",
                "consumes": [
                    "text/plain"
                ],
                "tags": [
                    "data"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Used to match provided data format to provider",
                        "name": "provider",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}